Fix some runtime errors (bug #13747)
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@35172 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/af/util/xp/ut_rand.cpp
@@ -596,7 +596,7 @@ random_r__ (struct UT_random_data *buf, UT_sint32 *result)
       UT_sint32 *end_ptr = buf->end_ptr;
       UT_sint32 val;
 
-      val = *fptr += *rptr;
+      val = *fptr += static_cast<long>(*rptr);
       /* Chucking least random bit.  */
       *result = (val >> 1) & 0x7fffffff;
       ++fptr;
